Reformatted the main game file
I reformatted the main game file so that we can actually have a level
two without 300 extra lines of code to isolate the two levels. The
buttons and labels are in different classes to separate and organize
them, so there is a general class for like the start game label and the
gam title label, and there is also a class for the player statistics
readouts and a class for each level.