PythonGameLib - Added in random damage values to the 'fist' weapon and added an experience multiplier that increases the weapon damage.
PTAG - Added transition to the hallway and sort of back to the room. There just needs to be a button bounds update for the hall and return to the room and there also needs to be a slight GUI update. Also some more story. (We are actually coming up to the first instance of combat. A very ambitious goal.)